uint32 -> int64 in databases

+++ ugc_service/ugc_db/init_migrations/001_create_tables.up.sql
@@ -1,8 +1,8 @@
 SET TIMEZONE="Europe/Moscow";
 
 CREATE TABLE posts (
-    post_id SERIAL NOT NULL,
-    author_id INTEGER NOT NULL,
+    post_id BIGSERIAL NOT NULL,
+    author_id BIGINT NOT NULL,
     content TEXT NOT NULL,
     create_timestamp TIMESTAMP WITH TIME ZONE NOT NULL,
     update_timestamp TIMESTAMP WITH TIME ZONE,
